Add MyThingsViewModel.kt Rename AccountFragment.kt to MyThingsFragment.kt
