+++ app/src/main/java/com/example/thingder/MainActivity.kt
@@ -23,7 +23,7 @@ class MainActivity : AppCompatActivity() {
             setOf(
                 R.id.navigation_main,
                 R.id.navigation_mine_liked_things,
-                R.id.navigation_account
+                R.id.navigation_my_things
             )
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
