--- app/src/main/java/com/example/thingder/fragments/account/AccountFragment.kt
@@ -1,21 +1,0 @@
-package com.example.thingder.fragments.account
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.View
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.thingder.R
-import com.example.thingder.databinding.FragmentAccountBinding
-
-class AccountFragment : Fragment(R.layout.fragment_account) {
-    private lateinit var binding: FragmentAccountBinding
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding = FragmentAccountBinding.bind(view)
-
-        binding.accountThingsRecycler.adapter = AccountThingsAdapter(emptyList())
-        binding.accountThingsRecycler.layoutManager = LinearLayoutManager(context,
-            LinearLayoutManager.VERTICAL, false)
-    }
-}
