+++ app/src/main/java/com/example/thingder/fragments/myThings/MyThingsAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.thingder.fragments.account
+package com.example.thingder.fragments.myThings
 
 import android.view.LayoutInflater
 import android.view.View
@@ -9,11 +9,11 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.thingder.R
 import com.example.thingder.domain.entities.Thing
 
-class AccountThingsAdapter(private val things: List<Thing>): RecyclerView.Adapter<ThingViewHolder>() {
+class MyThingsAdapter(private val things: List<Thing>): RecyclerView.Adapter<ThingViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ThingViewHolder {
         return ThingViewHolder(LayoutInflater.from(parent.context)
-            .inflate(R.layout.view_holder_account_things_item, parent, false))
+            .inflate(R.layout.view_holder_my_things_item, parent, false))
     }
 
     override fun onBindViewHolder(holder: ThingViewHolder, position: Int) {
@@ -25,8 +25,8 @@ class AccountThingsAdapter(private val things: List<Thing>): RecyclerView.Adapte
 }
 
 class ThingViewHolder(itemView: View): RecyclerView.ViewHolder(itemView) {
-    private val photo: ImageView = itemView.findViewById(R.id.view_holder_account_things_photo)
-    private val name: TextView = itemView.findViewById(R.id.view_holder_account_things_name)
+    private val photo: ImageView = itemView.findViewById(R.id.view_holder_my_things_photo)
+    private val name: TextView = itemView.findViewById(R.id.view_holder_my_things_name)
 
     fun onBind(thing: Thing) {
         name.text = thing.title
