+++ app/src/main/res/values/strings.xml
@@ -13,7 +13,7 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="nav_main_title">Main</string>
     <string name="nav_mine_liked_things_title">Mine Liked Things</string>
-    <string name="nav_account_title">Account</string>
+    <string name="nav_my_things_title">My Things</string>
 
     <string name="welcome_to_thingder">Welcome to Thingder</string>
     <string name="you_are_not_authorised">You are not authorised</string>
